ANDROID-10322 Removing colors
